backend/requirements.txt
# --- Flask & Extensions ---
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0

# --- Datenbank ---
PyMySQL==1.1.1

# --- Security & Utils ---
python-dotenv==1.0.1
argon2-cffi==23.1.0

# --- Server ---
gunicorn==21.2.0

# --- HTTP ---
requests==2.32.3

# --- SFTP / SSH ---
paramiko==3.4.0

# --- Hintergrundjobs & Redis ---
redis==5.0.8
rq==1.16.2

# --- Malware-Scanner Erweiterungen ---
yara-python==4.5.1
python-magic==0.4.27
